refactor: remove tokenizer object wrapper
