+++ src/propositional/utils/convert-prop-formula-to-expression.ts
@@ -1,6 +1,6 @@
 import { Operator } from "../../enums";
 import type { PropExpression, PropFormula } from "../../models";
-import { getOperatorGlyph } from "../../tokenizer/get-operator-glyph";
+import { getOperatorGlyph } from "../../tokenizer";
 import { createPropSymbol } from "../factory/create-prop-symbol";
 
 export function convertPropFormulaToExpression(
